+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.1.4 (Unreleased)
 - Make sure that states and reports are always dispatched to the partition `0`. This should prevent UI from not fully working when someone accidentally creates more partitions than expected.
+- Fix a bug where bootstrapping would create two initial states.
+- Introduce a /status page that can validate the setup and tell what is missing (#1318)
 
 ## 0.1.3 (2023-02-14)
 - Skip topics creation if web topics already exist (do not raise error)